\name{orderedPvalue-package}
\alias{orderedPvalue-package}
\alias{orderedPvalue}
\docType{package}
\title{
\packageTitle{orderedPvalue}
}
\description{
\packageDescription{orderedPvalue}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{orderedPvalue}
\packageIndices{orderedPvalue}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{orderedPvalue}

Maintainer: \packageMaintainer{orderedPvalue}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
